se hizo una mejora en el tipo de dato a long para guardar odenes de compras de materiales y otras mejoras para corregir implementacion de dispensaciones multiples

+++ src/main/java/exotic/app/planta/model/compras/OrdenCompraMateriales.java
@@ -44,9 +44,9 @@ public class OrdenCompraMateriales {
     @JsonManagedReference
     private List<ItemOrdenCompra> itemsOrdenCompra;
 
-    private int subTotal;
-    private int ivaCOP;
-    private int totalPagar;
+    private long subTotal;
+    private long ivaCOP;
+    private long totalPagar;
 
     /**
      * 0: credito
